requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
configparser==7.0.0
crispy-bootstrap5==0.7
cryptography==42.0.8
Django==4.0.10
django-bootstrap-v5==1.0.11
django-crispy-forms==2.0
django-crispy-formset-modal==0.11
django-extensions==3.1.5
django-minio-storage==0.3.10
djongo5==1.3.9
dnspython==2.6.1
fontawesomefree==6.5.1
idna==3.7
josepy==1.14.0
minio==6.0.2
mozilla-django-oidc==2.0.0
pycparser==2.22
pycryptodome==3.20.0
pymongo==3.12.3
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.2.4
typing_extensions==4.12.2
urllib3==2.2.1
